fix: Bookcomment 분리 및 더보기
